Introduced new basic nav-data infrastructure.

+++ de.knewcleus.fgfs/src/de/knewcleus/fgfs/Units.java
@@ -2,38 +2,38 @@ package de.knewcleus.fgfs;
 
 public class Units {
 	/* We're working in the SI-system */
-	public final static double M=1.0;
-	public final static double SEC=1.0;
-	public final static double KG=1.0;
+	public final static float M=1.0f;
+	public final static float SEC=1.0f;
+	public final static float KG=1.0f;
 
 	/* Distances */
-	public final static double FT=0.3048*M;
-	public final static double KM=1000.0*M;
-	public final static double NM=1852.0*M;
+	public final static float FT=0.3048f*M;
+	public final static float KM=1000.0f*M;
+	public final static float NM=1852.0f*M;
 	
 	/* Time */
-	public final static double MIN=60.0*SEC;
-	public final static double HOUR=60.0*MIN;
+	public final static float MIN=60.0f*SEC;
+	public final static float HOUR=60.0f*MIN;
 	
 	/* Frequency */
-	public final static double HZ=1.0;
-	public final static double KHZ=1.0E3*HZ;
-	public final static double MHZ=1.0E6*HZ;
+	public final static float HZ=1.0f;
+	public final static float KHZ=1.0E3f*HZ;
+	public final static float MHZ=1.0E6f*HZ;
 	
 	/* Angles */
-	public final static double DEG=1.0;
-	public final static double RAD=180.0/Math.PI*DEG;
-	public final static double FULLCIRCLE=360.0*DEG;
+	public final static double DEG=1.0f;
+	public final static double RAD=180.0f/Math.PI*DEG;
+	public final static double FULLCIRCLE=360.0f*DEG;
 
 	/* Velocities */
-	public final static double MPS=M/SEC;
-	public final static double FPM=FT/MIN;
-	public final static double KMH=KM/HOUR;
-	public final static double KNOTS=NM/HOUR;
+	public final static float MPS=M/SEC;
+	public final static float FPM=FT/MIN;
+	public final static float KMH=KM/HOUR;
+	public final static float KNOTS=NM/HOUR;
 	
 	/* Forces */
-	public final static double NEWTON=KG*M/(SEC*SEC);
+	public final static float NEWTON=KG*M/(SEC*SEC);
 
 	/* Constants */
-	public final static double g=9.81*NEWTON/KG;
+	public final static float g=9.81f*NEWTON/KG;
 }
